Exit thread waiting for connection on node removal
The loop that waits for an agent on mesos to connect assumes the
Computer is always valid. It is possible that a clean-up process
marked the Computer for removal and will loop forever (the agent
will never connect).

This change looks to see if the OfflineClause was set and if so,
no longer waits for a connection.